pyproject.toml
[project]
name = "hive"
version = "0.1.0"
description = "AI Tooling Omnimodal - Skywalking"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.1",
    "youtube-transcript-api>=1.2.3",
    "google-api-python-client>=2.187.0",
]

[project.optional-dependencies]
dev = []

[dependency-groups]
dev = []
